Fix build script

+++ build.sh
@@ -3,10 +3,13 @@
 # Build Axentro explorer
 
 MY_DIR=$(cd -P -- "$(dirname -- "$0")" && pwd -P)
-if [ -d $MY_DIR/[bin|lib] ]
-then
-    rm -rf $MY_DIR/[dist|lib]
-fi
+for dir in $MY_DIR/bin $MY_DIR/lib
+do
+    if [ -d $dir ]
+    then
+        rm -rf $dir
+    fi
+done
 
 if [ -x $MY_DIR/build_mintapp.sh ]
 then
@@ -18,7 +21,13 @@ then
     PREFIX=$HOME/opt/crystal-0.35.0-1/bin
 fi
 
-$PREFIX/shards install --production
+if [ -e $MY_DIR/shard.lock ]
+then
+    $PREFIX/shards update --production
+else
+    $PREFIX/shards install --production
+fi
+
 $PREFIX/shards build --no-debug --release --production
 
 cd "$OLDPWD"
